Move the DynamoConfig case class into the new module
