+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/dynamo/DynamoConfig.scala
@@ -1,3 +1,3 @@
-package uk.ac.wellcome.models.aws
+package uk.ac.wellcome.storage.dynamo
 
 case class DynamoConfig(table: String)
