+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/dynamo/VersionedDao.scala
@@ -8,7 +8,6 @@ import com.gu.scanamo.query.{KeyEquals, UniqueKey}
 import com.gu.scanamo.syntax.{attributeExists, not, _}
 import com.gu.scanamo.{DynamoFormat, Scanamo, Table}
 import com.twitter.inject.Logging
-import uk.ac.wellcome.models.aws.DynamoConfig
 import uk.ac.wellcome.type_classes.{IdGetter, VersionGetter, VersionUpdater}
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
